requirements.txt
Django==6.0.1
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
anthropic==0.40.0

# Logging
python-json-logger==2.0.7

# Code quality tools (for CI/CD)
coverage==7.4.0
flake8==7.0.0
black==24.1.1
isort==5.13.2

# Security scanning (for CI/CD)
safety==3.0.1
